refactor: extract navbar and add search results page
