+++ apps/trade-web/src/NavBar.css
